backend_crawling/crawler/requirements.txt
requests
schedule
psycopg2-binary
pandas
python-dotenv
smtpemail
